@@ -9,13 +9,12 @@
     {"name": "Thomas Weinert", "email": "thomas@weinert.info"}
   ],
   "require": {
-    "php": ">=5.6",
-    "carica/io": "dev-master"
+    "php": ">=7.1",
+    "carica/io": "^1.0"
   },
   "require-dev": {
-    "phake/phake": "1.*",
-    "carica/firmata": "dev-master",
-    "carica/gpio": "dev-master"
+    "carica/firmata": "^1.0",
+    "carica/gpio": "^0.0"
   },
   "autoload": {
     "psr-4": {
